Support mongrel2 buffered uploads
Mongrel2 can store the upload to disk for you.

You'll get 2 HTTP requests, one with HTTP_MONGREL2_UPLOAD_START and one
with also HTTP_MONGREL2_UPLOAD_DONE.

If you reply with an empty response, the upload will be cancelled. Use
`throw :async` on the upload start if you want to keep going on.